Mythril Enchant: Slightly increased buildup speed

+++ Core/ModPlayers/AccessoryFieldsPlayer.cs
@@ -106,7 +106,7 @@ namespace FargowiltasSouls.Core.ModPlayers
 
         public Item PlatinumEffect;
         public int PalladCounter;
-        public int MythrilTimer;
+        public float MythrilTimer;
         public int MythrilDelay;
         public int MythrilSoundCooldown;
         public int MythrilMaxTime => Player.HasEffect<MythrilEffect>() ? Player.ForceEffect<MythrilEffect>() ? 300 : 180 : 180;
